Extract ProblemExceptionNormalizer into AbstractProblemExceptionNormalizer and add Symfony 7 compatible method signature for ProblemExceptionNormalizer::normalize
