+++ src/ExceptionHandling/Normalizer/ProblemExceptionNormalizer.php
@@ -14,11 +14,8 @@ declare(strict_types=1);
 namespace Nijens\OpenapiBundle\ExceptionHandling\Normalizer;
 
 use Nijens\OpenapiBundle\ExceptionHandling\Exception\ProblemExceptionInterface;
-use Symfony\Component\HttpKernel\Exception\HttpExceptionInterface;
-use Symfony\Component\Serializer\Exception\InvalidArgumentException;
-use Symfony\Component\Serializer\Exception\LogicException;
+use Nijens\OpenapiBundle\NijensOpenapiBundle;
 use Symfony\Component\Serializer\Normalizer\NormalizerAwareInterface;
-use Symfony\Component\Serializer\Normalizer\NormalizerAwareTrait;
 use Throwable;
 
 /**
@@ -26,85 +23,23 @@ use Throwable;
  *
  * @author Niels Nijens <nijens.niels@gmail.com>
  */
-final class ProblemExceptionNormalizer implements NormalizerInterface, NormalizerAwareInterface
-{
-    use NormalizerAwareTrait;
-
-    private const ALREADY_CALLED = 'nijens_openapi.problem_exception_normalizer.already_called';
-
-    /**
-     * @var bool
-     */
-    private $debug;
-
-    public function __construct(bool $debug = false)
+if (NijensOpenapiBundle::getSymfonyVersion() < 70000) {
+    final class ProblemExceptionNormalizer extends AbstractProblemExceptionNormalizer implements NormalizerInterface, NormalizerAwareInterface
     {
-        $this->debug = $debug;
+        /**
+         * @return array
+         */
+        public function normalize($object, $format = null, array $context = []): float|int|bool|\ArrayObject|array|string|null
+        {
+            return $this->doNormalize($object, $format, $context);
+        }
     }
-
-    /**
-     * @return array
-     */
-    public function normalize($object, $format = null, array $context = [])
+} else {
+    final class ProblemExceptionNormalizer extends AbstractProblemExceptionNormalizer implements NormalizerInterface, NormalizerAwareInterface
     {
-        if ($object instanceof ProblemExceptionInterface === false) {
-            throw new InvalidArgumentException(sprintf('The object must implement "%s".', ProblemExceptionInterface::class));
+        public function normalize($object, $format = null, array $context = []): array
+        {
+            return $this->doNormalize($object, $format, $context);
         }
-
-        if (isset($context[self::ALREADY_CALLED])) {
-            throw new LogicException(sprintf('The normalizer "%s" can only be called once.', self::class));
-        }
-
-        $context[self::ALREADY_CALLED] = true;
-
-        $data = $this->normalizer->normalize($object, $format, $context);
-
-        $this->removeDetailsToPreventInformationDisclosure($object, $data);
-
-        return $this->unsetKeysWithNullValue($data);
-    }
-
-    public function supportsNormalization($data, $format = null, array $context = []): bool
-    {
-        if (isset($context[self::ALREADY_CALLED])) {
-            return false;
-        }
-
-        return $data instanceof ProblemExceptionInterface;
-    }
-
-    public function getSupportedTypes(?string $format): array
-    {
-        return [
-            ProblemExceptionInterface::class => false,
-        ];
-    }
-
-    private function removeDetailsToPreventInformationDisclosure(ProblemExceptionInterface $object, array &$data): void
-    {
-        if ($this->debug) {
-            return;
-        }
-
-        if ($object->getPrevious() === null || $object->getPrevious() instanceof HttpExceptionInterface) {
-            return;
-        }
-
-        unset($data['detail']);
-    }
-
-    private function unsetKeysWithNullValue(array $data): array
-    {
-        foreach ($data as $key => $value) {
-            if (is_array($value)) {
-                $data[$key] = $this->unsetKeysWithNullValue($value);
-            }
-
-            if ($value === null) {
-                unset($data[$key]);
-            }
-        }
-
-        return $data;
     }
 }
